Move quota dir
