+++ broker/lib/middleware.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const errors = require('../../common/errors');
 const logger = require('../../common/logger');
-const quota = require('../lib/quota');
+const quota = require('../../quota');
 const quotaManager = quota.quotaManager;
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
