+++ quota/QuotaManager.js
@@ -2,11 +2,11 @@
 
 const _ = require('lodash');
 const Promise = require('bluebird');
-const logger = require('../../../common/logger');
-const cf = require('../../../data-access-layer/cf');
-const catalog = require('../../../common/models/catalog');
-const config = require('../../../common/config');
-const CONST = require('../../../common/constants');
+const logger = require('../common/logger');
+const cf = require('../data-access-layer/cf');
+const catalog = require('../common/models/catalog');
+const config = require('../common/config');
+const CONST = require('../common/constants');
 
 class QuotaManager {
   constructor(quotaAPIClient) {
