+++ test/test_broker/quota.QuotaAPIClient.spec.js
@@ -4,8 +4,8 @@ const Promise = require('bluebird');
 const _ = require('lodash');
 const proxyquire = require('proxyquire');
 
-const QuotaAPIClient = proxyquire('../../broker/lib/quota/QuotaAPIClient', {
-  '../../../common/config': {
+const QuotaAPIClient = proxyquire('../../quota/QuotaAPIClient', {
+  '../common/config': {
     quota: {
       enabled: false,
       oauthDomain: 'sap-provisioning',
